Create README.md
Most people have a preference for movies of a similar genre. Movies within a genre often share common base parameters.
In this project, we will quantify the similarity of movies based on their plot summaries available on IMDb and Wikipedia, then separate them into groups, also known as clusters. We'll create a dendrogram to represent how closely the movies are related to each other.